% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{corporate_pal}
\alias{corporate_pal}
\title{A palette function from a corporate design}
\usage{
corporate_pal(design, type = c("qualitative", "sequential", "diverging"))
}
\arguments{
\item{design}{A \code{corporate_design} object.}

\item{type}{One of \code{"qualitative"}, \code{"sequential"} or \code{"diverging"}.}
}
\value{
A function \verb{function(n)} returning a character vector of colours.
}
\description{
Returns a function of \code{n} that produces \code{n} colours from the requested
palette. Qualitative palettes return the design colours directly (and
interpolate only if more colours than defined are requested); sequential and
diverging palettes interpolate between the design's endpoints.
}
\examples{
design <- corporate_design(
  system.file("designs/efk.yml", package = "CorporateDown")
)
corporate_pal(design, "qualitative")(3)

}
